Remove usued files
